Moved check for call log to ThankYou
